backend/requirements.txt
# FastAPI Backend Requirements for WorkZen HRMS
# Python 3.9+ required

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Security & Authentication
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1
python-jose[cryptography]>=3.3.0
cryptography>=41.0.7

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Environment
python-dotenv>=1.0.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Development (optional - uncomment if needed)
# pytest>=7.4.3
# pytest-asyncio>=0.21.1
# httpx>=0.25.2

# Code Quality (optional - uncomment if needed)
# black>=23.12.0
# flake8>=6.1.0
# mypy>=1.7.1
